Main Menu, Leaderboard Updates
Leaderboard testing needed.

+++ Wall2Wall0/app/src/main/java/com/example/jkt/wall2wall0/SpawnEvent.java
@@ -7,9 +7,11 @@ public class SpawnEvent {
 
     int enemy_type;
     int enemy_x_location;
-    int enemy_spawn_time;
+    float enemy_spawn_time;
 
-    public SpawnEvent(int type, int location, int time) {
+    // SpawnEvent class is used to hold all relevant attributes for each enemy spawn, this
+    // information will be used to create FallingEnemy objects in GameScreen class
+    public SpawnEvent(int type, int location, float time) {
         this.enemy_type = type;
         this.enemy_x_location = location;
         this.enemy_spawn_time = time;
